Operations refactoring
